backend

+++ src/main/java/infoz/domi/domain/EmployeRole.java
@@ -21,26 +21,24 @@ import javax.persistence.Table;
  */
 @Entity
 @Table(name = "EMPLOYE_ROLE")
-@NamedQueries({ @NamedQuery(name = "EmployeRole.findAll", query = "SELECT e FROM EmployeRole e") })
+@NamedQueries({
+    @NamedQuery(name = "EmployeRole.findAll", query = "SELECT e FROM EmployeRole e")})
 public class EmployeRole implements Serializable {
 
     private static final long serialVersionUID = 1L;
-
     @EmbeddedId
     protected EmployeRolePK employeRolePK;
-
     @Column(name = "CODE_NIVEAU")
-    private Long codeNiveau;
-
+    private Short codeNiveau;
     @JoinColumn(name = "MAT_EMP", referencedColumnName = "CODE_EMPLOYE", insertable = false, updatable = false)
     @ManyToOne(optional = false)
     private Employe employe;
-
     @JoinColumn(name = "CODE_ROLE", referencedColumnName = "CODE_ROLE", insertable = false, updatable = false)
     @ManyToOne(optional = false)
     private Role role;
 
-    public EmployeRole() {}
+    public EmployeRole() {
+    }
 
     public EmployeRole(EmployeRolePK employeRolePK) {
         this.employeRolePK = employeRolePK;
@@ -58,11 +56,11 @@ public class EmployeRole implements Serializable {
         this.employeRolePK = employeRolePK;
     }
 
-    public Long getCodeNiveau() {
+    public Short getCodeNiveau() {
         return codeNiveau;
     }
 
-    public void setCodeNiveau(Long codeNiveau) {
+    public void setCodeNiveau(Short codeNiveau) {
         this.codeNiveau = codeNiveau;
     }
 
@@ -96,10 +94,7 @@ public class EmployeRole implements Serializable {
             return false;
         }
         EmployeRole other = (EmployeRole) object;
-        if (
-            (this.employeRolePK == null && other.employeRolePK != null) ||
-            (this.employeRolePK != null && !this.employeRolePK.equals(other.employeRolePK))
-        ) {
+        if ((this.employeRolePK == null && other.employeRolePK != null) || (this.employeRolePK != null && !this.employeRolePK.equals(other.employeRolePK))) {
             return false;
         }
         return true;
@@ -109,4 +104,5 @@ public class EmployeRole implements Serializable {
     public String toString() {
         return "infoz.domi.domain.EmployeRole[ employeRolePK=" + employeRolePK + " ]";
     }
+    
 }
